// prodis-prodiswcl/src/app/modules/prodisapi/model/prospettoProvSede.ts
/*
* SPDX-FileCopyrightText: Copyright 2022 | Regione Piemonte
* SPDX-License-Identifier: EUPL-1.2
*/
/**
 * Prodis
 * API per il backend della suite di Prodis.
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { Comune } from './comune';
import { DatiProvinciali } from './datiProvinciali';


export interface ProspettoProvSede { 
    cap?: string;
    codUserAggiorn?: string;
    codUserInserim?: string;
    cognomeReferente?: string;
    comune?: Comune;
    dAggiorn?: Date;
    dInserim?: Date;
    datiProvinciali?: DatiProvinciali;
    email?: string;
    fax?: string;
    id?: number;
    indirizzo?: string;
    nomeReferente?: string;
    telefono?: string;
}